Network.AWS.Config

Description

Amazon Config is a fully managed service that provides you with an Amazon resource inventory, configuration history, and configuration change notifications to enable security and governance. With Amazon Config you can discover existing Amazon resources, export a complete inventory of your Amazon resources with all configuration details, and determine how a resource was configured at any point in time. These capabilities enable compliance auditing, security analysis, resource change tracking, and troubleshooting.

Documentation

module Network.AWS.Config.DeleteDeliveryChannel

module Network.AWS.Config.DeliverConfigSnapshot

module Network.AWS.Config.DescribeConfigurationRecorderStatus

module Network.AWS.Config.DescribeConfigurationRecorders

module Network.AWS.Config.DescribeDeliveryChannelStatus

module Network.AWS.Config.DescribeDeliveryChannels

module Network.AWS.Config.GetResourceConfigHistory

module Network.AWS.Config.PutConfigurationRecorder

module Network.AWS.Config.PutDeliveryChannel

module Network.AWS.Config.StartConfigurationRecorder

module Network.AWS.Config.StopConfigurationRecorder

module Network.AWS.Config.Types