Network.AWS.Config.DeleteConfigRule

Description

Deletes the specified AWS Config rule and all of its evaluation results.

AWS Config sets the state of a rule to DELETING until the deletion is complete. You cannot update a rule while it is in this state. If you make a PutConfigRule or DeleteConfigRule request for the rule, you will receive a ResourceInUseException .

You can check the state of a rule by using the DescribeConfigRules request.

Creating a Request

deleteConfigRule

Arguments

:: Text
dcrConfigRuleName

-> DeleteConfigRule

Creates a value of DeleteConfigRule with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dcrConfigRuleName - The name of the AWS Config rule that you want to delete.

data DeleteConfigRule

See: deleteConfigRule smart constructor.

Request Lenses

dcrConfigRuleName :: Lens' DeleteConfigRule Text

The name of the AWS Config rule that you want to delete.

Destructuring the Response

deleteConfigRuleResponse :: DeleteConfigRuleResponse

Creates a value of DeleteConfigRuleResponse with the minimum fields required to make a request.

data DeleteConfigRuleResponse

See: deleteConfigRuleResponse smart constructor.