Network.AWS.Config.DeleteConfigurationRecorder

Description

Deletes the configuration recorder.

After the configuration recorder is deleted, AWS Config will not record resource configuration changes until you create a new configuration recorder.

This action does not delete the configuration information that was previously recorded. You will be able to access the previously recorded information by using the GetResourceConfigHistory action, but you will not be able to access this information in the AWS Config console until you create a new configuration recorder.

Creating a Request

deleteConfigurationRecorder

Arguments

:: Text
dcrConfigurationRecorderName

-> DeleteConfigurationRecorder

Creates a value of DeleteConfigurationRecorder with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dcrConfigurationRecorderName - The name of the configuration recorder to be deleted. You can retrieve the name of your configuration recorder by using the DescribeConfigurationRecorders action.

data DeleteConfigurationRecorder

The request object for the DeleteConfigurationRecorder action.

See: deleteConfigurationRecorder smart constructor.

Request Lenses

dcrConfigurationRecorderName :: Lens' DeleteConfigurationRecorder Text

The name of the configuration recorder to be deleted. You can retrieve the name of your configuration recorder by using the DescribeConfigurationRecorders action.

Destructuring the Response

deleteConfigurationRecorderResponse :: DeleteConfigurationRecorderResponse

Creates a value of DeleteConfigurationRecorderResponse with the minimum fields required to make a request.

data DeleteConfigurationRecorderResponse

See: deleteConfigurationRecorderResponse smart constructor.