Network.AWS.Config.DeleteDeliveryChannel

Description

Deletes the specified delivery channel.

The delivery channel cannot be deleted if it is the only delivery channel and the configuration recorder is still running. To delete the delivery channel, stop the running configuration recorder using the StopConfigurationRecorder action.

http://docs.aws.amazon.com/config/latest/APIReference/API_DeleteDeliveryChannel.html

Request

data DeleteDeliveryChannel

Request constructor

deleteDeliveryChannel

Arguments

:: Text
ddcDeliveryChannelName

-> DeleteDeliveryChannel

DeleteDeliveryChannel constructor.

The fields accessible through corresponding lenses are:

ddcDeliveryChannelName :: Text

Request lenses

ddcDeliveryChannelName :: Lens' DeleteDeliveryChannel Text

The name of the delivery channel to delete.

Response

data DeleteDeliveryChannelResponse

Response constructor

deleteDeliveryChannelResponse :: DeleteDeliveryChannelResponse

DeleteDeliveryChannelResponse constructor.