Network.AWS.Config.DeliverConfigSnapshot

Description

Schedules delivery of a configuration snapshot to the Amazon S3 bucket in the specified delivery channel. After the delivery has started, AWS Config sends following notifications using an Amazon SNS topic that you have specified.

Notification of starting the delivery. Notification of delivery completed, if the delivery was successfully completed. Notification of delivery failure, if the delivery failed to complete.

http://docs.aws.amazon.com/config/latest/APIReference/API_DeliverConfigSnapshot.html

Request

data DeliverConfigSnapshot

Request constructor

deliverConfigSnapshot

Arguments

:: Text
dcsDeliveryChannelName

-> DeliverConfigSnapshot

DeliverConfigSnapshot constructor.

The fields accessible through corresponding lenses are:

dcsDeliveryChannelName :: Text

Request lenses

dcsDeliveryChannelName :: Lens' DeliverConfigSnapshot Text

The name of the delivery channel through which the snapshot is delivered.

Response

data DeliverConfigSnapshotResponse

Response constructor

deliverConfigSnapshotResponse :: DeliverConfigSnapshotResponse

DeliverConfigSnapshotResponse constructor.

The fields accessible through corresponding lenses are:

dcsrConfigSnapshotId :: Maybe Text

Response lenses

dcsrConfigSnapshotId :: Lens' DeliverConfigSnapshotResponse (Maybe Text)

The ID of the snapshot that is being created.