Network.AWS.Config.DescribeComplianceByConfigRule

Description

Indicates whether the specified AWS Config rules are compliant. If a rule is noncompliant, this action returns the number of AWS resources that do not comply with the rule.

A rule is compliant if all of the evaluated resources comply with it, and it is noncompliant if any of these resources do not comply.

If AWS Config has no current evaluation results for the rule, it returns INSUFFICIENT_DATA . This result might indicate one of the following conditions:

AWS Config has never invoked an evaluation for the rule. To check whether it has, use the DescribeConfigRuleEvaluationStatus action to get the LastSuccessfulInvocationTime and LastFailedInvocationTime .

The rule's AWS Lambda function is failing to send evaluation results to AWS Config. Verify that the role that you assigned to your configuration recorder includes the config:PutEvaluations permission. If the rule is a custom rule, verify that the AWS Lambda execution role includes the config:PutEvaluations permission.

The rule's AWS Lambda function has returned NOT_APPLICABLE for all evaluation results. This can occur if the resources were deleted or removed from the rule's scope.

This operation returns paginated results.

Creating a Request

describeComplianceByConfigRule :: DescribeComplianceByConfigRule

Creates a value of DescribeComplianceByConfigRule with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dcbcrConfigRuleNames - Specify one or more AWS Config rule names to filter the results by rule.

dcbcrComplianceTypes - Filters the results by compliance. The allowed values are COMPLIANT , NON_COMPLIANT , and INSUFFICIENT_DATA .

dcbcrNextToken - The NextToken string returned on a previous page that you use to get the next page of results in a paginated response.

data DescribeComplianceByConfigRule

See: describeComplianceByConfigRule smart constructor.

Request Lenses

dcbcrConfigRuleNames :: Lens' DescribeComplianceByConfigRule [Text]

Specify one or more AWS Config rule names to filter the results by rule.

dcbcrComplianceTypes :: Lens' DescribeComplianceByConfigRule [ComplianceType]

Filters the results by compliance. The allowed values are COMPLIANT , NON_COMPLIANT , and INSUFFICIENT_DATA .

dcbcrNextToken :: Lens' DescribeComplianceByConfigRule (Maybe Text)

The NextToken string returned on a previous page that you use to get the next page of results in a paginated response.

Destructuring the Response

describeComplianceByConfigRuleResponse

Arguments

:: Int
dcbcrrsResponseStatus

-> DescribeComplianceByConfigRuleResponse

Creates a value of DescribeComplianceByConfigRuleResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dcbcrrsComplianceByConfigRules - Indicates whether each of the specified AWS Config rules is compliant.

dcbcrrsNextToken - The string that you use in a subsequent request to get the next page of results in a paginated response.

dcbcrrsResponseStatus - -- | The response status code.

data DescribeComplianceByConfigRuleResponse

See: describeComplianceByConfigRuleResponse smart constructor.

Response Lenses

dcbcrrsComplianceByConfigRules :: Lens' DescribeComplianceByConfigRuleResponse [ComplianceByConfigRule]

Indicates whether each of the specified AWS Config rules is compliant.

dcbcrrsNextToken :: Lens' DescribeComplianceByConfigRuleResponse (Maybe Text)

The string that you use in a subsequent request to get the next page of results in a paginated response.

dcbcrrsResponseStatus :: Lens' DescribeComplianceByConfigRuleResponse Int

- | The response status code.