Network.AWS.Config.DescribeComplianceByResource

Description

Indicates whether the specified AWS resources are compliant. If a resource is noncompliant, this action returns the number of AWS Config rules that the resource does not comply with.

A resource is compliant if it complies with all the AWS Config rules that evaluate it. It is noncompliant if it does not comply with one or more of these rules.

If AWS Config has no current evaluation results for the resource, it returns INSUFFICIENT_DATA . This result might indicate one of the following conditions about the rules that evaluate the resource:

AWS Config has never invoked an evaluation for the rule. To check whether it has, use the DescribeConfigRuleEvaluationStatus action to get the LastSuccessfulInvocationTime and LastFailedInvocationTime .

The rule's AWS Lambda function is failing to send evaluation results to AWS Config. Verify that the role that you assigned to your configuration recorder includes the config:PutEvaluations permission. If the rule is a custom rule, verify that the AWS Lambda execution role includes the config:PutEvaluations permission.

The rule's AWS Lambda function has returned NOT_APPLICABLE for all evaluation results. This can occur if the resources were deleted or removed from the rule's scope.

This operation returns paginated results.

Creating a Request

describeComplianceByResource :: DescribeComplianceByResource

Creates a value of DescribeComplianceByResource with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dcbrResourceId - The ID of the AWS resource for which you want compliance information. You can specify only one resource ID. If you specify a resource ID, you must also specify a type for ResourceType .

dcbrResourceType - The types of AWS resources for which you want compliance information; for example, AWS::EC2::Instance . For this action, you can specify that the resource type is an AWS account by specifying AWS::::Account .

dcbrComplianceTypes - Filters the results by compliance. The allowed values are COMPLIANT , NON_COMPLIANT , and INSUFFICIENT_DATA .

dcbrNextToken - The NextToken string returned on a previous page that you use to get the next page of results in a paginated response.

dcbrLimit - The maximum number of evaluation results returned on each page. The default is 10. You cannot specify a limit greater than 100. If you specify 0, AWS Config uses the default.

data DescribeComplianceByResource

See: describeComplianceByResource smart constructor.

Request Lenses

dcbrResourceId :: Lens' DescribeComplianceByResource (Maybe Text)

The ID of the AWS resource for which you want compliance information. You can specify only one resource ID. If you specify a resource ID, you must also specify a type for ResourceType .

dcbrResourceType :: Lens' DescribeComplianceByResource (Maybe Text)

The types of AWS resources for which you want compliance information; for example, AWS::EC2::Instance . For this action, you can specify that the resource type is an AWS account by specifying AWS::::Account .

dcbrComplianceTypes :: Lens' DescribeComplianceByResource [ComplianceType]

Filters the results by compliance. The allowed values are COMPLIANT , NON_COMPLIANT , and INSUFFICIENT_DATA .

dcbrNextToken :: Lens' DescribeComplianceByResource (Maybe Text)

The NextToken string returned on a previous page that you use to get the next page of results in a paginated response.

dcbrLimit :: Lens' DescribeComplianceByResource (Maybe Natural)

The maximum number of evaluation results returned on each page. The default is 10. You cannot specify a limit greater than 100. If you specify 0, AWS Config uses the default.

Destructuring the Response

describeComplianceByResourceResponse

Arguments

:: Int
dcbrrsResponseStatus

-> DescribeComplianceByResourceResponse

Creates a value of DescribeComplianceByResourceResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dcbrrsComplianceByResources - Indicates whether the specified AWS resource complies with all of the AWS Config rules that evaluate it.

dcbrrsNextToken - The string that you use in a subsequent request to get the next page of results in a paginated response.

dcbrrsResponseStatus - -- | The response status code.

data DescribeComplianceByResourceResponse

See: describeComplianceByResourceResponse smart constructor.

Response Lenses

dcbrrsComplianceByResources :: Lens' DescribeComplianceByResourceResponse [ComplianceByResource]

Indicates whether the specified AWS resource complies with all of the AWS Config rules that evaluate it.

dcbrrsNextToken :: Lens' DescribeComplianceByResourceResponse (Maybe Text)

The string that you use in a subsequent request to get the next page of results in a paginated response.

dcbrrsResponseStatus :: Lens' DescribeComplianceByResourceResponse Int

- | The response status code.