Network.AWS.Config.DescribeConfigRuleEvaluationStatus

Description

Returns status information for each of your AWS managed Config rules. The status includes information such as the last time AWS Config invoked the rule, the last time AWS Config failed to invoke the rule, and the related error for the last failure.

Creating a Request

describeConfigRuleEvaluationStatus :: DescribeConfigRuleEvaluationStatus

Creates a value of DescribeConfigRuleEvaluationStatus with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dcresConfigRuleNames - The name of the AWS managed Config rules for which you want status information. If you do not specify any names, AWS Config returns status information for all AWS managed Config rules that you use.

dcresNextToken - The NextToken string returned on a previous page that you use to get the next page of results in a paginated response.

dcresLimit - The number of rule evaluation results that you want returned. This parameter is required if the rule limit for your account is more than the default of 50 rules. For more information about requesting a rule limit increase, see AWS Config Limits in the AWS General Reference Guide .

data DescribeConfigRuleEvaluationStatus

See: describeConfigRuleEvaluationStatus smart constructor.

Request Lenses

dcresConfigRuleNames :: Lens' DescribeConfigRuleEvaluationStatus [Text]

The name of the AWS managed Config rules for which you want status information. If you do not specify any names, AWS Config returns status information for all AWS managed Config rules that you use.

dcresNextToken :: Lens' DescribeConfigRuleEvaluationStatus (Maybe Text)

The NextToken string returned on a previous page that you use to get the next page of results in a paginated response.

dcresLimit :: Lens' DescribeConfigRuleEvaluationStatus (Maybe Natural)

The number of rule evaluation results that you want returned. This parameter is required if the rule limit for your account is more than the default of 50 rules. For more information about requesting a rule limit increase, see AWS Config Limits in the AWS General Reference Guide .

Destructuring the Response

describeConfigRuleEvaluationStatusResponse

Arguments

:: Int
dcresrsResponseStatus

-> DescribeConfigRuleEvaluationStatusResponse

Creates a value of DescribeConfigRuleEvaluationStatusResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dcresrsConfigRulesEvaluationStatus - Status information about your AWS managed Config rules.

dcresrsNextToken - The string that you use in a subsequent request to get the next page of results in a paginated response.

dcresrsResponseStatus - -- | The response status code.

data DescribeConfigRuleEvaluationStatusResponse

See: describeConfigRuleEvaluationStatusResponse smart constructor.

Response Lenses

dcresrsConfigRulesEvaluationStatus :: Lens' DescribeConfigRuleEvaluationStatusResponse [ConfigRuleEvaluationStatus]

Status information about your AWS managed Config rules.

dcresrsNextToken :: Lens' DescribeConfigRuleEvaluationStatusResponse (Maybe Text)

The string that you use in a subsequent request to get the next page of results in a paginated response.

dcresrsResponseStatus :: Lens' DescribeConfigRuleEvaluationStatusResponse Int

- | The response status code.