Network.AWS.Config.DescribeConfigRules

Description

Returns details about your AWS Config rules.

This operation returns paginated results.

Creating a Request

describeConfigRules :: DescribeConfigRules

Creates a value of DescribeConfigRules with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dcrConfigRuleNames - The names of the AWS Config rules for which you want details. If you do not specify any names, AWS Config returns details for all your rules.

dcrNextToken - The NextToken string returned on a previous page that you use to get the next page of results in a paginated response.

data DescribeConfigRules

See: describeConfigRules smart constructor.

Request Lenses

dcrConfigRuleNames :: Lens' DescribeConfigRules [Text]

The names of the AWS Config rules for which you want details. If you do not specify any names, AWS Config returns details for all your rules.

dcrNextToken :: Lens' DescribeConfigRules (Maybe Text)

The NextToken string returned on a previous page that you use to get the next page of results in a paginated response.

Destructuring the Response

describeConfigRulesResponse

Arguments

:: Int
dcrrsResponseStatus

-> DescribeConfigRulesResponse

Creates a value of DescribeConfigRulesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dcrrsConfigRules - The details about your AWS Config rules.

dcrrsNextToken - The string that you use in a subsequent request to get the next page of results in a paginated response.

dcrrsResponseStatus - -- | The response status code.

data DescribeConfigRulesResponse

See: describeConfigRulesResponse smart constructor.

Response Lenses

dcrrsConfigRules :: Lens' DescribeConfigRulesResponse [ConfigRule]

The details about your AWS Config rules.

dcrrsNextToken :: Lens' DescribeConfigRulesResponse (Maybe Text)

The string that you use in a subsequent request to get the next page of results in a paginated response.

dcrrsResponseStatus :: Lens' DescribeConfigRulesResponse Int

- | The response status code.