Network.AWS.Config.DescribeConfigurationRecorderStatus

Description

Returns the current status of the specified configuration recorder. If a configuration recorder is not specified, this action returns the status of all configuration recorder associated with the account.

http://docs.aws.amazon.com/config/latest/APIReference/API_DescribeConfigurationRecorderStatus.html

Request

data DescribeConfigurationRecorderStatus

Request constructor

describeConfigurationRecorderStatus :: DescribeConfigurationRecorderStatus

DescribeConfigurationRecorderStatus constructor.

The fields accessible through corresponding lenses are:

dcrsConfigurationRecorderNames :: [Text]

Request lenses

dcrsConfigurationRecorderNames :: Lens' DescribeConfigurationRecorderStatus [Text]

The name(s) of the configuration recorder. If the name is not specified, the action returns the current status of all the configuration recorders associated with the account.

Response

data DescribeConfigurationRecorderStatusResponse

Response constructor

describeConfigurationRecorderStatusResponse :: DescribeConfigurationRecorderStatusResponse

DescribeConfigurationRecorderStatusResponse constructor.

The fields accessible through corresponding lenses are:

dcrsrConfigurationRecordersStatus :: [ConfigurationRecorderStatus]

Response lenses

dcrsrConfigurationRecordersStatus :: Lens' DescribeConfigurationRecorderStatusResponse [ConfigurationRecorderStatus]

A list that contains status of the specified recorders.