Network.AWS.Config.DescribeConfigurationRecorders

Description

Returns the details for the specified configuration recorders. If the configuration recorder is not specified, this action returns the details for all configuration recorders associated with the account.

Creating a Request

describeConfigurationRecorders :: DescribeConfigurationRecorders

Creates a value of DescribeConfigurationRecorders with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dcrConfigurationRecorderNames - A list of configuration recorder names.

data DescribeConfigurationRecorders

The input for the DescribeConfigurationRecorders action.

See: describeConfigurationRecorders smart constructor.

Request Lenses

dcrConfigurationRecorderNames :: Lens' DescribeConfigurationRecorders [Text]

A list of configuration recorder names.

Destructuring the Response

describeConfigurationRecordersResponse

Arguments

:: Int
drsResponseStatus

-> DescribeConfigurationRecordersResponse

Creates a value of DescribeConfigurationRecordersResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

drsConfigurationRecorders - A list that contains the descriptions of the specified configuration recorders.

drsResponseStatus - -- | The response status code.

data DescribeConfigurationRecordersResponse

The output for the DescribeConfigurationRecorders action.

See: describeConfigurationRecordersResponse smart constructor.

Response Lenses

drsConfigurationRecorders :: Lens' DescribeConfigurationRecordersResponse [ConfigurationRecorder]

A list that contains the descriptions of the specified configuration recorders.

drsResponseStatus :: Lens' DescribeConfigurationRecordersResponse Int

- | The response status code.