Network.AWS.Config.DescribeDeliveryChannelStatus

Description

Returns the current status of the specified delivery channel. If a delivery channel is not specified, this action returns the current status of all delivery channels associated with the account.

http://docs.aws.amazon.com/config/latest/APIReference/API_DescribeDeliveryChannelStatus.html

Request

data DescribeDeliveryChannelStatus

Request constructor

describeDeliveryChannelStatus :: DescribeDeliveryChannelStatus

DescribeDeliveryChannelStatus constructor.

The fields accessible through corresponding lenses are:

ddcsDeliveryChannelNames :: [Text]

Request lenses

ddcsDeliveryChannelNames :: Lens' DescribeDeliveryChannelStatus [Text]

A list of delivery channel names.

Response

data DescribeDeliveryChannelStatusResponse

Response constructor

describeDeliveryChannelStatusResponse :: DescribeDeliveryChannelStatusResponse

DescribeDeliveryChannelStatusResponse constructor.

The fields accessible through corresponding lenses are:

ddcsrDeliveryChannelsStatus :: [DeliveryChannelStatus]

Response lenses

ddcsrDeliveryChannelsStatus :: Lens' DescribeDeliveryChannelStatusResponse [DeliveryChannelStatus]

A list that contains the status of a specified delivery channel.