Network.AWS.Config.GetComplianceDetailsByResource

Description

Returns the evaluation results for the specified AWS resource. The results indicate which AWS Config rules were used to evaluate the resource, when each rule was last used, and whether the resource complies with each rule.

This operation returns paginated results.

Creating a Request

getComplianceDetailsByResource

Arguments

:: Text
gcdbrResourceType

-> Text
gcdbrResourceId

-> GetComplianceDetailsByResource

Creates a value of GetComplianceDetailsByResource with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gcdbrComplianceTypes - Filters the results by compliance. The allowed values are COMPLIANT , NON_COMPLIANT , and NOT_APPLICABLE .

gcdbrNextToken - The NextToken string returned on a previous page that you use to get the next page of results in a paginated response.

gcdbrResourceType - The type of the AWS resource for which you want compliance information.

gcdbrResourceId - The ID of the AWS resource for which you want compliance information.

data GetComplianceDetailsByResource

See: getComplianceDetailsByResource smart constructor.

Request Lenses

gcdbrComplianceTypes :: Lens' GetComplianceDetailsByResource [ComplianceType]

Filters the results by compliance. The allowed values are COMPLIANT , NON_COMPLIANT , and NOT_APPLICABLE .

gcdbrNextToken :: Lens' GetComplianceDetailsByResource (Maybe Text)

The NextToken string returned on a previous page that you use to get the next page of results in a paginated response.

gcdbrResourceType :: Lens' GetComplianceDetailsByResource Text

The type of the AWS resource for which you want compliance information.

gcdbrResourceId :: Lens' GetComplianceDetailsByResource Text

The ID of the AWS resource for which you want compliance information.

Destructuring the Response

getComplianceDetailsByResourceResponse

Arguments

:: Int
gcdbrrsResponseStatus

-> GetComplianceDetailsByResourceResponse

Creates a value of GetComplianceDetailsByResourceResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gcdbrrsEvaluationResults - Indicates whether the specified AWS resource complies each AWS Config rule.

gcdbrrsNextToken - The string that you use in a subsequent request to get the next page of results in a paginated response.

gcdbrrsResponseStatus - -- | The response status code.

data GetComplianceDetailsByResourceResponse

See: getComplianceDetailsByResourceResponse smart constructor.

Response Lenses

gcdbrrsEvaluationResults :: Lens' GetComplianceDetailsByResourceResponse [EvaluationResult]

Indicates whether the specified AWS resource complies each AWS Config rule.

gcdbrrsNextToken :: Lens' GetComplianceDetailsByResourceResponse (Maybe Text)

The string that you use in a subsequent request to get the next page of results in a paginated response.

gcdbrrsResponseStatus :: Lens' GetComplianceDetailsByResourceResponse Int

- | The response status code.