Network.AWS.Config.GetComplianceSummaryByConfigRule

Description

Returns the number of AWS Config rules that are compliant and noncompliant, up to a maximum of 25 for each.

Creating a Request

getComplianceSummaryByConfigRule :: GetComplianceSummaryByConfigRule

Creates a value of GetComplianceSummaryByConfigRule with the minimum fields required to make a request.

data GetComplianceSummaryByConfigRule

See: getComplianceSummaryByConfigRule smart constructor.

Destructuring the Response

getComplianceSummaryByConfigRuleResponse

Arguments

:: Int
gcsbcrrsResponseStatus

-> GetComplianceSummaryByConfigRuleResponse

Creates a value of GetComplianceSummaryByConfigRuleResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gcsbcrrsComplianceSummary - The number of AWS Config rules that are compliant and the number that are noncompliant, up to a maximum of 25 for each.

gcsbcrrsResponseStatus - -- | The response status code.

data GetComplianceSummaryByConfigRuleResponse

See: getComplianceSummaryByConfigRuleResponse smart constructor.

Response Lenses

gcsbcrrsComplianceSummary :: Lens' GetComplianceSummaryByConfigRuleResponse (Maybe ComplianceSummary)

The number of AWS Config rules that are compliant and the number that are noncompliant, up to a maximum of 25 for each.

gcsbcrrsResponseStatus :: Lens' GetComplianceSummaryByConfigRuleResponse Int

- | The response status code.