Network.AWS.Config.GetResourceConfigHistory

Description

Returns a list of configuration items for the specified resource. The list contains details about each state of the resource during the specified time interval. You can specify a limit on the number of results returned on the page. If a limit is specified, a nextToken is returned as part of the result that you can use to continue this request.

http://docs.aws.amazon.com/config/latest/APIReference/API_GetResourceConfigHistory.html

Request

data GetResourceConfigHistory

Request constructor

getResourceConfigHistory

Arguments

:: ResourceType
grchResourceType

-> Text
grchResourceId

-> GetResourceConfigHistory

GetResourceConfigHistory constructor.

The fields accessible through corresponding lenses are:

grchChronologicalOrder :: Maybe ChronologicalOrder

grchEarlierTime :: Maybe UTCTime

grchLaterTime :: Maybe UTCTime

grchLimit :: Maybe Natural

grchNextToken :: Maybe Text

grchResourceId :: Text

grchResourceType :: ResourceType

Request lenses

grchChronologicalOrder :: Lens' GetResourceConfigHistory (Maybe ChronologicalOrder)

The chronological order for configuration items listed. By default the results are listed in reverse chronological order.

grchEarlierTime :: Lens' GetResourceConfigHistory (Maybe UTCTime)

The time stamp that indicates an earlier time. If not specified, the action returns paginated results that contain configuration items that start from when the first configuration item was recorded.

grchLaterTime :: Lens' GetResourceConfigHistory (Maybe UTCTime)

The time stamp that indicates a later time. If not specified, current time is taken.

grchLimit :: Lens' GetResourceConfigHistory (Maybe Natural)

The maximum number of configuration items returned in each page. The default is 10. You cannot specify a limit greater than 100.

grchNextToken :: Lens' GetResourceConfigHistory (Maybe Text)

An optional parameter used for pagination of the results.

grchResourceId :: Lens' GetResourceConfigHistory Text

The ID of the resource (for example., 'sg-xxxxxx').

grchResourceType :: Lens' GetResourceConfigHistory ResourceType

The resource type.

Response

data GetResourceConfigHistoryResponse

Response constructor

getResourceConfigHistoryResponse :: GetResourceConfigHistoryResponse

GetResourceConfigHistoryResponse constructor.

The fields accessible through corresponding lenses are:

grchrConfigurationItems :: [ConfigurationItem]

grchrNextToken :: Maybe Text

Response lenses

grchrConfigurationItems :: Lens' GetResourceConfigHistoryResponse [ConfigurationItem]

A list that contains the configuration history of one or more resources.

grchrNextToken :: Lens' GetResourceConfigHistoryResponse (Maybe Text)

A token used for pagination of results.