Network.AWS.Config.PutConfigurationRecorder

Description

Creates a new configuration recorder to record the resource configurations.

You can use this action to change the role (roleARN) of an existing recorder. To change the role, call the action on the existing configuration recorder and specify a role.

http://docs.aws.amazon.com/config/latest/APIReference/API_PutConfigurationRecorder.html

Request

data PutConfigurationRecorder

Request constructor

putConfigurationRecorder

Arguments

:: ConfigurationRecorder
pcrConfigurationRecorder

-> PutConfigurationRecorder

PutConfigurationRecorder constructor.

The fields accessible through corresponding lenses are:

pcrConfigurationRecorder :: ConfigurationRecorder

Request lenses

pcrConfigurationRecorder :: Lens' PutConfigurationRecorder ConfigurationRecorder

The configuration recorder object that records each configuration change made to the resources.

Response

data PutConfigurationRecorderResponse

Response constructor

putConfigurationRecorderResponse :: PutConfigurationRecorderResponse

PutConfigurationRecorderResponse constructor.