Network.AWS.Config.PutDeliveryChannel

Description

Creates a new delivery channel object to deliver the configuration information to an Amazon S3 bucket, and to an Amazon SNS topic.

You can use this action to change the Amazon S3 bucket or an Amazon SNS topic of the existing delivery channel. To change the Amazon S3 bucket or an Amazon SNS topic, call this action and specify the changed values for the S3 bucket and the SNS topic. If you specify a different value for either the S3 bucket or the SNS topic, this action will keep the existing value for the parameter that is not changed.

http://docs.aws.amazon.com/config/latest/APIReference/API_PutDeliveryChannel.html

Request

data PutDeliveryChannel

Request constructor

putDeliveryChannel

Arguments

:: DeliveryChannel
pdcDeliveryChannel

-> PutDeliveryChannel

PutDeliveryChannel constructor.

The fields accessible through corresponding lenses are:

pdcDeliveryChannel :: DeliveryChannel

Request lenses

pdcDeliveryChannel :: Lens' PutDeliveryChannel DeliveryChannel

The configuration delivery channel object that delivers the configuration information to an Amazon S3 bucket, and to an Amazon SNS topic.

Response

data PutDeliveryChannelResponse

Response constructor

putDeliveryChannelResponse :: PutDeliveryChannelResponse

PutDeliveryChannelResponse constructor.