Network.AWS.Config.StartConfigurationRecorder

Description

Starts recording configurations of the AWS resources you have selected to record in your AWS account.

You must have created at least one delivery channel to successfully start the configuration recorder.

Creating a Request

startConfigurationRecorder

Arguments

:: Text
sConfigurationRecorderName

-> StartConfigurationRecorder

Creates a value of StartConfigurationRecorder with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

sConfigurationRecorderName - The name of the recorder object that records each configuration change made to the resources.

data StartConfigurationRecorder

The input for the StartConfigurationRecorder action.

See: startConfigurationRecorder smart constructor.

Request Lenses

sConfigurationRecorderName :: Lens' StartConfigurationRecorder Text

The name of the recorder object that records each configuration change made to the resources.

Destructuring the Response

startConfigurationRecorderResponse :: StartConfigurationRecorderResponse

Creates a value of StartConfigurationRecorderResponse with the minimum fields required to make a request.

data StartConfigurationRecorderResponse

See: startConfigurationRecorderResponse smart constructor.