Network.AWS.Config.StopConfigurationRecorder

Description

Stops recording configurations of all the resources associated with the account.

http://docs.aws.amazon.com/config/latest/APIReference/API_StopConfigurationRecorder.html

Request

data StopConfigurationRecorder

Request constructor

stopConfigurationRecorder

Arguments

:: Text
scrConfigurationRecorderName

-> StopConfigurationRecorder

StopConfigurationRecorder constructor.

The fields accessible through corresponding lenses are:

scrConfigurationRecorderName :: Text

Request lenses

scrConfigurationRecorderName :: Lens' StopConfigurationRecorder Text

The name of the recorder object that records each configuration change made to the resources.

Response

data StopConfigurationRecorderResponse

Response constructor

stopConfigurationRecorderResponse :: StopConfigurationRecorderResponse

StopConfigurationRecorderResponse constructor.