Network.AWS.Config.Types

Service

data Config

Version 2014-11-12 of the Amazon Config service.

Error

data JSONError :: *

ConfigExportDeliveryInfo

data ConfigExportDeliveryInfo

configExportDeliveryInfo :: ConfigExportDeliveryInfo

ConfigExportDeliveryInfo constructor.

The fields accessible through corresponding lenses are:

cediLastAttemptTime :: Maybe UTCTime

cediLastErrorCode :: Maybe Text

cediLastErrorMessage :: Maybe Text

cediLastStatus :: Maybe DeliveryStatus

cediLastSuccessfulTime :: Maybe UTCTime

cediLastAttemptTime :: Lens' ConfigExportDeliveryInfo (Maybe UTCTime)

The time of the last attempted delivery.

cediLastErrorCode :: Lens' ConfigExportDeliveryInfo (Maybe Text)

The error code from the last attempted delivery.

cediLastErrorMessage :: Lens' ConfigExportDeliveryInfo (Maybe Text)

The error message from the last attempted delivery.

cediLastStatus :: Lens' ConfigExportDeliveryInfo (Maybe DeliveryStatus)

Status of the last attempted delivery.

cediLastSuccessfulTime :: Lens' ConfigExportDeliveryInfo (Maybe UTCTime)

The time of the last successful delivery.

ConfigStreamDeliveryInfo

data ConfigStreamDeliveryInfo

configStreamDeliveryInfo :: ConfigStreamDeliveryInfo

ConfigStreamDeliveryInfo constructor.

The fields accessible through corresponding lenses are:

csdiLastErrorCode :: Maybe Text

csdiLastErrorMessage :: Maybe Text

csdiLastStatus :: Maybe DeliveryStatus

csdiLastStatusChangeTime :: Maybe UTCTime

csdiLastErrorCode :: Lens' ConfigStreamDeliveryInfo (Maybe Text)

The error code from the last attempted delivery.

csdiLastErrorMessage :: Lens' ConfigStreamDeliveryInfo (Maybe Text)

The error message from the last attempted delivery.

csdiLastStatus :: Lens' ConfigStreamDeliveryInfo (Maybe DeliveryStatus)

Status of the last attempted delivery.

csdiLastStatusChangeTime :: Lens' ConfigStreamDeliveryInfo (Maybe UTCTime)

The time from the last status change.

Relationship

data Relationship

relationship :: Relationship

Relationship constructor.

The fields accessible through corresponding lenses are:

rRelationshipName :: Maybe Text

rResourceId :: Maybe Text

rResourceType :: Maybe ResourceType

rRelationshipName :: Lens' Relationship (Maybe Text)

The name of the related resource.

rResourceId :: Lens' Relationship (Maybe Text)

The resource ID of the related resource (for example, 'sg-xxxxxx'.

rResourceType :: Lens' Relationship (Maybe ResourceType)

The resource type of the related resource.

DeliveryChannel

data DeliveryChannel

deliveryChannel :: DeliveryChannel

DeliveryChannel constructor.

The fields accessible through corresponding lenses are:

dcName :: Maybe Text

dcS3BucketName :: Maybe Text

dcS3KeyPrefix :: Maybe Text

dcSnsTopicARN :: Maybe Text

dcName :: Lens' DeliveryChannel (Maybe Text)

The name of the delivery channel. By default, AWS Config automatically assigns the name "default" when creating the delivery channel. You cannot change the assigned name.

dcS3BucketName :: Lens' DeliveryChannel (Maybe Text)

The name of the Amazon S3 bucket used to store configuration history for the delivery channel.

dcS3KeyPrefix :: Lens' DeliveryChannel (Maybe Text)

The prefix for the specified Amazon S3 bucket.

dcSnsTopicARN :: Lens' DeliveryChannel (Maybe Text)

The Amazon Resource Name (ARN) of the IAM role used for accessing the Amazon S3 bucket and the Amazon SNS topic.

ChronologicalOrder

data ChronologicalOrder

Constructors

Forward
Forward

Reverse
Reverse

ResourceType

data ResourceType

Constructors

AWSCloudTrailTrail
AWS::CloudTrail::Trail

AWSEC2CustomerGateway
AWS::EC2::CustomerGateway

AWSEC2EIP
AWS::EC2::EIP

AWSEC2Instance
AWS::EC2::Instance

AWSEC2InternetGateway
AWS::EC2::InternetGateway

AWSEC2NetworkAcl
AWS::EC2::NetworkAcl

AWSEC2NetworkInterface
AWS::EC2::NetworkInterface

AWSEC2RouteTable
AWS::EC2::RouteTable

AWSEC2SecurityGroup
AWS::EC2::SecurityGroup

AWSEC2Subnet
AWS::EC2::Subnet

AWSEC2VPC
AWS::EC2::VPC

AWSEC2VPNConnection
AWS::EC2::VPNConnection

AWSEC2VPNGateway
AWS::EC2::VPNGateway

AWSEC2Volume
AWS::EC2::Volume

ConfigurationItem

data ConfigurationItem

configurationItem :: ConfigurationItem

ConfigurationItem constructor.

The fields accessible through corresponding lenses are:

ciAccountId :: Maybe Text

ciArn :: Maybe Text

ciAvailabilityZone :: Maybe Text

ciConfiguration :: Maybe Text

ciConfigurationItemCaptureTime :: Maybe UTCTime

ciConfigurationItemMD5Hash :: Maybe Text

ciConfigurationItemStatus :: Maybe ConfigurationItemStatus

ciConfigurationStateId :: Maybe Text

ciRelatedEvents :: [Text]

ciRelationships :: [Relationship]

ciResourceCreationTime :: Maybe UTCTime

ciResourceId :: Maybe Text

ciResourceType :: Maybe ResourceType

ciTags :: HashMap Text Text

ciVersion :: Maybe Text

ciAccountId :: Lens' ConfigurationItem (Maybe Text)

The 12 digit AWS account ID associated with the resource.

ciArn :: Lens' ConfigurationItem (Maybe Text)

The Amazon Resource Name (ARN) of the resource.

ciAvailabilityZone :: Lens' ConfigurationItem (Maybe Text)

The Availability Zone associated with the resource.

ciConfiguration :: Lens' ConfigurationItem (Maybe Text)

The description of the resource configuration.

ciConfigurationItemCaptureTime :: Lens' ConfigurationItem (Maybe UTCTime)

The time when the configuration recording was initiated.

ciConfigurationItemMD5Hash :: Lens' ConfigurationItem (Maybe Text)

Unique MD5 hash that represents the configuration item's state.

You can use MD5 hash to compare the states of two or more configuration items that are associated with the same resource.

ciConfigurationItemStatus :: Lens' ConfigurationItem (Maybe ConfigurationItemStatus)

The configuration item status.

ciConfigurationStateId :: Lens' ConfigurationItem (Maybe Text)

An identifier that indicates the ordering of the configuration items of a resource.

ciRelatedEvents :: Lens' ConfigurationItem [Text]

A list of CloudTrail event IDs.

A populated field indicates that the current configuration was initiated by the events recorded in the CloudTrail log. For more information about CloudTrail, see What is AWS CloudTrail?.

An empty field indicates that the current configuration was not initiated by any event.

ciRelationships :: Lens' ConfigurationItem [Relationship]

A list of related AWS resources.

ciResourceCreationTime :: Lens' ConfigurationItem (Maybe UTCTime)

The time stamp when the resource was created.

ciResourceId :: Lens' ConfigurationItem (Maybe Text)

The ID of the resource (for example., 'sg-xxxxxx').

ciResourceType :: Lens' ConfigurationItem (Maybe ResourceType)

The type of AWS resource.

ciTags :: Lens' ConfigurationItem (HashMap Text Text)

A mapping of key value tags associated with the resource.

ciVersion :: Lens' ConfigurationItem (Maybe Text)

The version number of the resource configuration.

DeliveryStatus

data DeliveryStatus

Constructors

Failure
Failure

Success
Success

DeliveryChannelStatus

data DeliveryChannelStatus

deliveryChannelStatus :: DeliveryChannelStatus

DeliveryChannelStatus constructor.

The fields accessible through corresponding lenses are:

dcsConfigHistoryDeliveryInfo :: Maybe ConfigExportDeliveryInfo

dcsConfigSnapshotDeliveryInfo :: Maybe ConfigExportDeliveryInfo

dcsConfigStreamDeliveryInfo :: Maybe ConfigStreamDeliveryInfo

dcsName :: Maybe Text

dcsConfigHistoryDeliveryInfo :: Lens' DeliveryChannelStatus (Maybe ConfigExportDeliveryInfo)

A list that contains the status of the delivery of the configuration history to the specified Amazon S3 bucket.

dcsConfigSnapshotDeliveryInfo :: Lens' DeliveryChannelStatus (Maybe ConfigExportDeliveryInfo)

A list containing the status of the delivery of the snapshot to the specified Amazon S3 bucket.

dcsConfigStreamDeliveryInfo :: Lens' DeliveryChannelStatus (Maybe ConfigStreamDeliveryInfo)

A list containing the status of the delivery of the configuration stream notification to the specified Amazon SNS topic.

dcsName :: Lens' DeliveryChannelStatus (Maybe Text)

The name of the delivery channel.

ConfigurationRecorderStatus

data ConfigurationRecorderStatus

configurationRecorderStatus :: ConfigurationRecorderStatus

ConfigurationRecorderStatus constructor.

The fields accessible through corresponding lenses are:

crsLastErrorCode :: Maybe Text

crsLastErrorMessage :: Maybe Text

crsLastStartTime :: Maybe UTCTime

crsLastStatus :: Maybe RecorderStatus

crsLastStatusChangeTime :: Maybe UTCTime

crsLastStopTime :: Maybe UTCTime

crsName :: Maybe Text

crsRecording :: Maybe Bool

crsLastErrorCode :: Lens' ConfigurationRecorderStatus (Maybe Text)

The error code indicating that the recording failed.

crsLastErrorMessage :: Lens' ConfigurationRecorderStatus (Maybe Text)

The message indicating that the recording failed due to an error.

crsLastStartTime :: Lens' ConfigurationRecorderStatus (Maybe UTCTime)

The time the recorder was last started.

crsLastStatus :: Lens' ConfigurationRecorderStatus (Maybe RecorderStatus)

The last (previous) status of the recorder.

crsLastStatusChangeTime :: Lens' ConfigurationRecorderStatus (Maybe UTCTime)

The time when the status was last changed.

crsLastStopTime :: Lens' ConfigurationRecorderStatus (Maybe UTCTime)

The time the recorder was last stopped.

crsName :: Lens' ConfigurationRecorderStatus (Maybe Text)

The name of the configuration recorder.

crsRecording :: Lens' ConfigurationRecorderStatus (Maybe Bool)

Specifies whether the recorder is currently recording or not.

ConfigurationItemStatus

data ConfigurationItemStatus

Constructors

Deleted
Deleted

Discovered
Discovered

Failed
Failed

Ok
Ok

ConfigurationRecorder

data ConfigurationRecorder

configurationRecorder :: ConfigurationRecorder

ConfigurationRecorder constructor.

The fields accessible through corresponding lenses are:

crName :: Maybe Text

crRoleARN :: Maybe Text

crName :: Lens' ConfigurationRecorder (Maybe Text)

The name of the recorder. By default, AWS Config automatically assigns the name "default" when creating the configuration recorder. You cannot change the assigned name.

crRoleARN :: Lens' ConfigurationRecorder (Maybe Text)

Amazon Resource Name (ARN) of the IAM role used to describe the AWS resources associated with the account.

RecorderStatus

data RecorderStatus

Constructors

RSFailure
Failure

RSPending
Pending

RSSuccess
Success