Network.AWS.Config.DeliverConfigSnapshot

Description

Schedules delivery of a configuration snapshot to the Amazon S3 bucket in the specified delivery channel. After the delivery has started, AWS Config sends following notifications using an Amazon SNS topic that you have specified.

Notification of starting the delivery.

Notification of delivery completed, if the delivery was successfully completed.

Notification of delivery failure, if the delivery failed to complete.

Creating a Request

deliverConfigSnapshot

Arguments

:: Text
dcsDeliveryChannelName

-> DeliverConfigSnapshot

Creates a value of DeliverConfigSnapshot with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dcsDeliveryChannelName - The name of the delivery channel through which the snapshot is delivered.

data DeliverConfigSnapshot

The input for the DeliverConfigSnapshot action.

See: deliverConfigSnapshot smart constructor.

Request Lenses

dcsDeliveryChannelName :: Lens' DeliverConfigSnapshot Text

The name of the delivery channel through which the snapshot is delivered.

Destructuring the Response

deliverConfigSnapshotResponse

Arguments

:: Int
dcsrsResponseStatus

-> DeliverConfigSnapshotResponse

Creates a value of DeliverConfigSnapshotResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dcsrsConfigSnapshotId - The ID of the snapshot that is being created.

dcsrsResponseStatus - -- | The response status code.

data DeliverConfigSnapshotResponse

The output for the DeliverConfigSnapshot action in JSON format.

See: deliverConfigSnapshotResponse smart constructor.

Response Lenses

dcsrsConfigSnapshotId :: Lens' DeliverConfigSnapshotResponse (Maybe Text)

The ID of the snapshot that is being created.

dcsrsResponseStatus :: Lens' DeliverConfigSnapshotResponse Int

- | The response status code.