Network.AWS.Config.DescribeConfigurationRecorderStatus

Description

Returns the current status of the specified configuration recorder. If a configuration recorder is not specified, this action returns the status of all configuration recorder associated with the account.

Creating a Request

describeConfigurationRecorderStatus :: DescribeConfigurationRecorderStatus

Creates a value of DescribeConfigurationRecorderStatus with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dcrsConfigurationRecorderNames - The name(s) of the configuration recorder. If the name is not specified, the action returns the current status of all the configuration recorders associated with the account.

data DescribeConfigurationRecorderStatus

The input for the DescribeConfigurationRecorderStatus action.

See: describeConfigurationRecorderStatus smart constructor.

Request Lenses

dcrsConfigurationRecorderNames :: Lens' DescribeConfigurationRecorderStatus [Text]

The name(s) of the configuration recorder. If the name is not specified, the action returns the current status of all the configuration recorders associated with the account.

Destructuring the Response

describeConfigurationRecorderStatusResponse

Arguments

:: Int
dcrsrsResponseStatus

-> DescribeConfigurationRecorderStatusResponse

Creates a value of DescribeConfigurationRecorderStatusResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dcrsrsConfigurationRecordersStatus - A list that contains status of the specified recorders.

dcrsrsResponseStatus - -- | The response status code.

data DescribeConfigurationRecorderStatusResponse

The output for the DescribeConfigurationRecorderStatus action in JSON format.

See: describeConfigurationRecorderStatusResponse smart constructor.

Response Lenses

dcrsrsConfigurationRecordersStatus :: Lens' DescribeConfigurationRecorderStatusResponse [ConfigurationRecorderStatus]

A list that contains status of the specified recorders.

dcrsrsResponseStatus :: Lens' DescribeConfigurationRecorderStatusResponse Int

- | The response status code.