Network.AWS.Config.PutConfigurationRecorder

Description

Creates a new configuration recorder to record the selected resource configurations.

You can use this action to change the role roleARN and/or the recordingGroup of an existing recorder. To change the role, call the action on the existing configuration recorder and specify a role.

Creating a Request

putConfigurationRecorder

Arguments

:: ConfigurationRecorder
pcrConfigurationRecorder

-> PutConfigurationRecorder

Creates a value of PutConfigurationRecorder with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

pcrConfigurationRecorder - The configuration recorder object that records each configuration change made to the resources.

data PutConfigurationRecorder

The input for the PutConfigurationRecorder action.

See: putConfigurationRecorder smart constructor.

Request Lenses

pcrConfigurationRecorder :: Lens' PutConfigurationRecorder ConfigurationRecorder

The configuration recorder object that records each configuration change made to the resources.

Destructuring the Response

putConfigurationRecorderResponse :: PutConfigurationRecorderResponse

Creates a value of PutConfigurationRecorderResponse with the minimum fields required to make a request.

data PutConfigurationRecorderResponse

See: putConfigurationRecorderResponse smart constructor.